requirements.txt
-e git+https://github.com/leovnoliveira/ai-agents-quiz-jornada-de-dados.git@78f5aee0922b707adc9874e7f7b367a8427a1766#egg=ai_agents_quiz_jornada_de_dados
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
blobfile==3.0.0
bs4==0.0.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fairscale==0.4.13
filelock==3.18.0
fire==0.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
llama_models==0.1.5
llama_stack==0.1.7
llama_stack_client==0.1.7
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
protobuf==5.29.4
pyaml==25.1.0
pyarrow==19.0.1
pycryptodomex==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
replicate==1.0.4
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
setuptools==77.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
streamlit==1.43.2
sympy==1.13.1
tenacity==9.0.0
termcolor==2.5.0
tiktoken==0.9.0
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
